refactor: split workspace.py into separate cosmotech, k8s, and superset modules
